Dailydividend/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
feedparser>=6.0.10
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.0.267
langchain-core>=0.1.0
langchain-groq>=0.1.0
yfinance>=0.2.36
celery>=5.3.6
redis>=5.0.1
twilio>=8.10.0
gevent>=24.2.1
python-dateutil>=2.8.2
beautifulsoup4>=4.12.2
pytz>=2023.3
lxml>=4.9.3